fix(auth): move partner resolution out of the shared constants module
